Calendar.vue script seperated to composable
